Make sure int is provided to DecimalType convert instead of int

+++ src/main/java/org/openhab/binding/honeywellhome/client/api/pojo/ChangeableValues.java
@@ -2,8 +2,8 @@ package org.openhab.binding.honeywellhome.client.api.pojo;
 
 public class ChangeableValues {
     public String mode;
-    public int heatSetpoint;
-    public int coolSetpoint;
+    public long heatSetpoint;
+    public long coolSetpoint;
     public String thermostatSetpointStatus;
     public String heatCoolMode;
